Working on targeting edits.
I was hoping to do this all in Alpine.js. But, I'm starting to think this is going to get messy. I'm considering switching over to Angular 18 - would be fun to learn more about the latest Angular anyway. Though, I might just get a v1 done in Alpine so that I can think things through without a build-step.